Split content item request specs into live and draft.
